+++ src/address.rs
@@ -136,3 +136,11 @@ pub enum AddressType {
     LEPublic = 1,
     LERandom = 2,
 }
+
+#[repr(u32)]
+#[derive(Debug, Copy, Clone, FromPrimitive, ToPrimitive)]
+pub enum Protocol {
+    L2CAP = bluez_sys::BTPROTO_L2CAP,
+    HCI = bluez_sys::BTPROTO_HCI,
+    RFCOMM = bluez_sys::BTPROTO_RFCOMM,
+}
