+++ src/management/client/query.rs
@@ -1,7 +1,7 @@
 use crate::AddressType;
 use std::collections::HashMap;
 
-use crate::management::interface::controller::ControllerInfoExt;
+use crate::management::interface::ControllerInfoExt;
 use crate::util::BufExtBlueZ;
 
 use super::*;
