+++ src/management/mod.rs
@@ -1,6 +1,8 @@
-pub mod client;
+mod client;
+mod stream;
 pub mod interface;
 pub mod result;
-pub mod stream;
 
 pub use result::{Error, Result};
+pub use stream::ManagementStream;
+pub use client::*;
